requirements.txt
# Browser Automation Agent - Dependencies
# Python 3.11

# Core browser automation
playwright==1.49.1
playwright-stealth==1.0.6
undetected-playwright

# Environment configuration
python-dotenv==1.0.1

# LLM integration
ollama>=0.4.0
httpx>=0.27.0

# Vector database for memory
chromadb>=0.4.0

# Note: Run setup.py to install and configure everything
